Rename contants.ts -> constants.ts
