+++ frontend/src/games/dot-count-game.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import { delay } from '@/utils/timing';
-import { BACKEND_URL } from '@/utils/contants';
+import { BACKEND_URL } from '@/utils/constants';
 import { Vec2 } from 'src/utils/vec2';
 
 type Dot = {
