+++ frontend/src/games/memory-game-with-numbers.tsx
@@ -1,5 +1,5 @@
 import { useState, useEffect } from 'react';
-import { BACKEND_URL } from '@/utils/contants';
+import { BACKEND_URL } from '@/utils/constants';
 
 const MemoryGameWithNumbers = () => {
   const [grid, setGrid] = useState<(number | null)[]>([]);
